+++ modules/02_inventory.ps1
@@ -1,5 +1,5 @@
 ﻿###############################################################################
-# Maintenance Toolkit 3.0.6.2 - Modulo inventario
+# Maintenance Toolkit 3.7.0 - Modulo inventario
 ###############################################################################
 
 . "$PSScriptRoot\00_common.ps1"